Add size check signature
Since it's (typically) uncommon for large binaries to be stored in
repositories. If a large file has been commited there is a chance it was
mistaken and could be an accidental information disclosure.

Since the size check doesn't conform to the pattern of other signatures
it means the signature validation functions have been relaxed.